pyproject.toml
# SPDX-FileCopyrightText: 2024 Max Mehl <https://mehl.mx>
#
# SPDX-License-Identifier: GPL-3.0-only

[tool.poetry]
name = "inwx-dns-recordmaster"
version = "0.1.1"
description = "Manage DNS nameserver records of INWX domains via YAML files and API requests. Lightweight, version-control ready"
authors = ["Max Mehl <mail@mehl.mx>"]
readme = "README.md"
packages = [{ include = "recordmaster" }]

[tool.poetry.scripts]
inwx-dnsrm = "recordmaster.main:main"

[tool.poetry.dependencies]
python = "^3.11"
inwx-domrobot = "^3.1.0"
pyyaml = "^6.0.1"
platformdirs = "^4.2.0"
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"
pylama = "^8.4.1"
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"
types-pyyaml = "^6.0.12.12"
types-toml = "^0.10.8.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# FORMATTING settings
[tool.isort]
profile = "black"

[tool.black]
line-length = 100

# MYPY settings
[tool.mypy]
files = ["recordmaster/*.py"]
